+++ assets/javascripts/kitten/components/form/text-input-with-limit/text-input-with-limit.stories.js
@@ -1,12 +1,24 @@
 import React from 'react'
-import { boolean, text } from '@storybook/addon-knobs'
+import { boolean, select, number } from '@storybook/addon-knobs'
 import { TextInputWithLimit } from './index'
+
+const variantOptions = {
+  Andromeda: 'andromeda',
+  Orion: 'orion',
+}
+
+const tagOptions = {
+  Input: 'input',
+  Textarea: 'textarea',
+}
 
 export const Default = () => (
   <TextInputWithLimit
     disabled={boolean('Disabled', false)}
     tiny={boolean('Tiny', false)}
-    tag={text('input', 'input')}
+    tag={select('Tag', tagOptions, 'input')}
+    variant={select('Variant', variantOptions, 'andromeda')}
+    limit={number('Limit', 80)}
   />
 )