+++ assets/javascripts/kitten/components/form/checkbox/checkbox.stories.js
@@ -16,11 +16,36 @@ export const Default = () => (
   <Grid>
     <GridCol offset="1" col="8">
       <Checkbox
-        id={text('id', 'checkbox_id')}
         onLabelClick={action('label-click')}
         indeterminate={boolean('indeterminate', false)}
+        id={text('id', 'checkbox_id')}
+        disabled={boolean('disabled', false)}
       >
         {text('text', 'Checkbox Label')}
+      </Checkbox>
+    </GridCol>
+    <GridCol offset="1" col="8">
+      <Checkbox
+        id="checkbox_id_2"
+        indeterminate={boolean('indeterminate', false)}
+        disabled={boolean('disabled', false)}
+      >
+        This label has a <a href="#link">link</a>.
+      </Checkbox>
+    </GridCol>
+  </Grid>
+)
+
+export const WithContent = () => (
+  <Grid>
+    <GridCol offset="1" col="8">
+      <Checkbox
+        indeterminate={boolean('indeterminate', false)}
+        id={text('id', 'checkbox_id')}
+        disabled={boolean('disabled', false)}
+        label={text('label', 'Checkbox Label')}
+      >
+        {text('content', 'Checkbox content')}
       </Checkbox>
     </GridCol>
   </Grid>