+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- Feature: Update `RewardCard` component.
 
 ## [2.95.3] - 2020-11-05