+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
-- Feature: Update `RewardCard` component.
+
+## [2.96.0] - 2020-11-12
+Feature: 
+- Update `RewardCard` component.
 
 ## [2.95.3] - 2020-11-05